% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_95ci.R
\name{mean_95ci}
\alias{mean_95ci}
\title{Display result in "mean (lci, uci)"}
\usage{
mean_95ci(mean, lci, uci, dig)
}
\arguments{
\item{mean}{Mean value}

\item{lci}{Lower 95\% CI}

\item{uci}{Upper 95\% CI}

\item{dig}{Desired number of digits for display}
}
\value{
A string in the form of "mean (lci, uci)"
}
\description{
Display result in "mean (lci, uci)"
}
\examples{
mean_95ci(1.02, 0.86, 1.21, 2)
}
