% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Display results in percentages}
\usage{
pct(x)
}
\arguments{
\item{x}{numeric of any length}
}
\value{
The same numeric with decimal place moved 2 digits to the right.
}
\description{
Display results in percentages
}
\examples{
x <- 0.8612
pct(x)
}
